+++ src/onegov.town/onegov/town/utils.py
@@ -141,11 +141,19 @@ class AllocationEventInfo(object):
 
     @property
     def event_actions(self):
-        yield Link(
-            _("Reserve"),
-            self.request.link(self.allocation, name='reservieren'),
-            classes=('new-reservation', )
-        )
+
+        if self.availability == 0:
+            yield Link(
+                _("Unavailable"),
+                '#',
+                classes=('new-reservation', 'disabled')
+            )
+        else:
+            yield Link(
+                _("Reserve"),
+                self.request.link(self.allocation, name='reservieren'),
+                classes=('new-reservation', )
+            )
 
         if self.request.is_logged_in:
             yield Link(